mise en place de l'interaction avec le frontend

+++ src/core/interactions/transaction/deleteTransactionUseCase.ts
@@ -1,4 +1,5 @@
 import { NotFoundError } from "../../errors/notFoundError";
+import { RecordRepository } from "../repositories/recordRepository";
 import { TransactionRepository } from "../repositories/transactionRepository";
 
 interface IDeleteTransactionUseCase {
@@ -13,10 +14,12 @@ export interface IDeleteTransactoinUseCaseResponse {
 export class DeleteTransactionUseCase implements IDeleteTransactionUseCase {
     private repository: TransactionRepository;
     private presenter: IDeleteTransactoinUseCaseResponse;
+    private record_repo: RecordRepository;
 
-    constructor(repo: TransactionRepository, presenter: IDeleteTransactoinUseCaseResponse) {
+    constructor(repo: TransactionRepository, record_repo: RecordRepository, presenter: IDeleteTransactoinUseCaseResponse) {
         this.repository = repo;
         this.presenter = presenter;
+        this.record_repo = record_repo;
     }
     
     async execute(id: string): Promise<void> {
@@ -24,6 +27,12 @@ export class DeleteTransactionUseCase implements IDeleteTransactionUseCase {
             let transaction = await this.repository.get(id);
             if (transaction == null) {
                 throw new NotFoundError('Transaction not found');
+            }
+
+            let is_deleted_record = await this.record_repo.get(transaction.record.id);
+
+            if (is_deleted_record) {
+                throw new NotFoundError('Error while deleting record');
             }
 
             let is_deleted = await this.repository.delete(id);
